requirements.txt
folioclient>=1.0.7
PyYAML>=6.0.2
